+++ client/tests/end_to_end.rs
@@ -59,14 +59,14 @@ async fn register_project() {
     let register_project_message = random_register_project_message(org_id.clone(), checkpoint_id);
     let project_name = register_project_message.project_name.clone();
     let random_fee = random_balance();
-    let tx_applied = submit_ok_with_fee(
+    let tx_included = submit_ok_with_fee(
         &client,
         &author,
         register_project_message.clone(),
         random_fee,
     )
     .await;
-    assert_eq!(tx_applied.result, Ok(()));
+    assert_eq!(tx_included.result, Ok(()));
 
     let project = client
         .get_project(project_name.clone(), org_id.clone())
@@ -82,7 +82,7 @@ async fn register_project() {
     assert_eq!(project.metadata.clone(), register_project_message.metadata);
 
     assert_eq!(
-        tx_applied.events[0],
+        tx_included.events[0],
         RegistryEvent::ProjectRegistered(project_name.clone(), org_id.clone()).into()
     );
 
@@ -130,14 +130,14 @@ async fn register_org() {
 
     let register_org_message = random_register_org_message();
     let random_fee = random_balance();
-    let tx_applied =
+    let tx_included =
         submit_ok_with_fee(&client, &author, register_org_message.clone(), random_fee).await;
 
     assert_eq!(
-        tx_applied.events[0],
+        tx_included.events[0],
         RegistryEvent::OrgRegistered(register_org_message.org_id.clone()).into()
     );
-    assert_eq!(tx_applied.result, Ok(()));
+    assert_eq!(tx_included.result, Ok(()));
 
     let opt_org = client
         .get_org(register_org_message.org_id.clone())
@@ -165,11 +165,10 @@ async fn register_user() {
     let author = ed25519::Pair::from_string("//Alice", None).unwrap();
 
     let register_user_message = random_register_user_message();
-    let tx_applied = submit_ok(&client, &author, register_user_message.clone()).await;
-    assert_eq!(tx_applied.result, Ok(()));
+    let tx_included = submit_ok(&client, &author, register_user_message.clone()).await;
 
     assert_eq!(
-        tx_applied.events[0],
+        tx_included.events[0],
         RegistryEvent::UserRegistered(register_user_message.user_id.clone()).into(),
     );