+++ runtime-tests/tests/checkpoing_setting.rs
@@ -141,7 +141,7 @@ async fn set_checkpoint_without_permission() {
 
     let (bad_actor, _) = key_pair_with_associated_user(&client).await;
 
-    let tx_applied = submit_ok(
+    let tx_included = submit_ok(
         &client,
         &bad_actor,
         message::SetCheckpoint {
@@ -158,7 +158,7 @@ async fn set_checkpoint_without_permission() {
         .unwrap()
         .unwrap();
     assert_eq!(
-        tx_applied.result,
+        tx_included.result,
         Err(RegistryError::InsufficientSenderPermissions.into())
     );
     assert_eq!(updated_project.current_cp, project.current_cp.clone());
@@ -175,7 +175,7 @@ async fn fail_to_set_nonexistent_checkpoint() {
     let project_name = project.name.clone();
     let garbage = CheckpointId::random();
 
-    let tx_applied = submit_ok(
+    let tx_included = submit_ok(
         &client,
         &author,
         message::SetCheckpoint {
@@ -187,7 +187,7 @@ async fn fail_to_set_nonexistent_checkpoint() {
     .await;
 
     assert_eq!(
-        tx_applied.result,
+        tx_included.result,
         Err(RegistryError::InexistentCheckpointId.into())
     );
     let updated_project = client